Refactor authentication logic and update profile screen; replace login screen with new implementation
